refactor(docs): satisfy source layout limits
